prepared react routes for other html
