+++ client/src/App.js
@@ -1,21 +1,25 @@
 import React, { Component } from "react";
-import "./App.css";
-import Navbar from "./components/navigation/navigation/Navbar";
-import "../node_modules/bootstrap/dist/js/bootstrap.bundle";
-import "../node_modules/bootstrap/dist/css/bootstrap.min.css";
+import LandingPageWrapper from './components/landingPage/LandingPageWrapper';
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
-class App extends Component {
-  render() {
-    return (
-      <Router>
-        <Switch>
-          <Navbar /> {/*Add some paths here when others get their pages done*/}
-        </Switch>
-        <body class="bg-color"></body>
-      </Router>
-    );
-  }
+import "./App.css";
+import "../node_modules/bootstrap/dist/js/bootstrap.bundle";
+import "../node_modules/bootstrap/dist/css/bootstrap.min.css";
+
+
+  class App extends Component {
+
+    render() {
+      return (
+
+       <Router>
+          <Switch>
+            <LandingPageWrapper path ="/" exact component={LandingPageWrapper}/> {/*Add some paths here when others get their pages done*/}
+          </Switch>
+         <body class="bg-color"></body>
+        </Router>
+      );
+   }
 }
 
 export default App;
