+++ client/src/components/landingPage/LandingPageWrapper.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import Navbar from '../navigation/Navbar';
+
+
+class LandingPageWrapper extends Component {
+  render(){
+    return(
+      <div>
+        <Navbar />
+      </div>
+    )
+  }
+}
+
+export default LandingPageWrapper;
