--- client/src/components/navigation/navigation/Sidebar.css
@@ -1,26 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Roboto&display=swap');
-
-.custom-navbar {
- background-color: #39304A;
- width: 15rem;
- height: 100vh;
- box-shadow: 14px 10px 15px rgba(0, 0, 0, 0.24);
- position: fixed;
- font-family: 'Roboto', sans-serif;
-}
-
-hr {
-  display: block;
-  margin-top: 0.5em;
-  margin-bottom: 0.5em;
-  margin-left: auto;
-  margin-right: auto;
-  border-style: inset;
-  border-width: 2px;
-  background-color: white;
-  
-}
-
-button {
- border: none;
-}
